# cmd/api/docs/swagger.yaml
definitions:
...
  domain.Admins:
    properties:
      id:
        type: integer
      password:
        type: string
      username:
        type: string
    type: object
...
  domain.Login:
    properties:
      email:
        maxLength: 50
        minLength: 3
        type: string
      password:
        maxLength: 12
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
...
  domain.Users:
    properties:
      block_Status:
        type: boolean
      city:
        type: string
      country:
        type: string
      created_At:
        type: string
      email:
        maxLength: 50
        minLength: 3
        type: string
      first_name:
        maxLength: 12
        minLength: 3
        type: string
      gender:
        maxLength: 8
        minLength: 4
        type: string
      last_name:
        maxLength: 12
        minLength: 1
        type: string
      password:
        maxLength: 12
        minLength: 6
        type: string
      phone:
        maxLength: 12
        minLength: 3
        type: string
      profile_Pic:
        type: string
      updated_At:
        type: string
      user_Id:
        type: integer
      verification:
        type: boolean
    required:
    - email
    - first_name
    - gender
    - last_name
    - password
    - phone
    type: object
  response.Response:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      status:
        type: boolean
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server  server. You can visit the GitHub repository
    at https://github.com/prasanth-pn/clean-code-architecture-ecommerce
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Go + Gin ecommerce API
  version: "1.0"
paths:
...
  /admin/login:
    post:
      operationId: AdminLogin for Admin
      parameters:
      - description: AdminLogin
        in: body
        name: AdminLogin
        required: true
        schema:
          $ref: '#/definitions/domain.Admins'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      summary: Admin Login for Admin
      tags:
      - ADMINLOGIN
...
  /admin/register:
    post:
      operationId: AdminRegister for Admin
      parameters:
      - description: AdminRegister
        in: body
        name: AdminRegister
        required: true
        schema:
          $ref: '#/definitions/domain.Admins'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      summary: Register for Admin
      tags:
      - Admin
...
  /user/login:
    post:
      operationId: user login authentication
      parameters:
      - description: userLogin
        in: body
        name: userLogin
        required: true
        schema:
          $ref: '#/definitions/domain.Login'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      summary: Login for user
      tags:
      - USERTAUTHENTICATION
...
  /user/register:
    post:
      operationId: Register for user
      parameters:
      - description: Register
        in: body
        name: Register
        required: true
        schema:
          $ref: '#/definitions/domain.Users'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      summary: Register for user
      tags:
      - USERTAUTHENTICATION
  /user/send/verificationmail:
    post:
      operationId: SendUserMail
      parameters:
      - description: email
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      summary: Send EmailOtp for User
      tags:
      - USERVERIFICATION
...
  /user/verify/otp:
    get:
      operationId: VerifyUserOtp for authentication
      parameters:
      - description: email
        in: query
        name: email
        required: true
        type: string
      - description: code
        in: query
        name: code
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      summary: VerifyUserOtp for User
      tags:
      - USERVERIFICATION
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"